requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-filter==1.1.0
django-rest==0.0.1
djangorestframework==3.8.2
idna==2.7
Markdown==2.6.11
psycopg2==2.7.5
pytz==2018.4
requests==2.19.1
urllib3==1.23

